+++ app/javascript/lca/components/characters/charms/CharmEditor.jsx
@@ -6,18 +6,29 @@ import Button from 'material-ui/Button'
 import Grid from 'material-ui/Grid'
 import Typography from 'material-ui/Typography'
 import ContentAddCircle from 'material-ui-icons/AddCircle'
+import Filter from 'material-ui-icons/FilterList'
 
 import CharmFields from './CharmFields.jsx'
+import CharmFilter from './CharmFilter.jsx'
 import SpellFields from './SpellFields.jsx'
 import {
   updateCharm, createCharm, destroyCharm,
   updateSpell, createSpell, destroySpell,
 } from '../../../ducks/actions.js'
+import {
+  getSpecificCharacter, getMeritsForCharacter, getNativeCharmsForCharacter,
+  getMartialArtsCharmsForCharacter, getEvocationsForCharacter,
+  getSpellsForCharacter, getSpiritCharmsForCharacter,
+  getAllAbilitiesWithCharmsForCharacter,
+} from '../../../selectors/'
 
 class CharmEditor extends React.Component {
   constructor(props) {
     super(props)
 
+    this.state = { charmFilter: '' }
+
+    this.setFilter = this.setFilter.bind(this)
     this.handleUpdate = this.handleUpdate.bind(this)
     this.handleAddNative = this.handleAddNative.bind(this)
     this.handleAddMA = this.handleAddMA.bind(this)
@@ -27,6 +38,10 @@ class CharmEditor extends React.Component {
     this.handleUpdateSpell = this.handleUpdateSpell.bind(this)
     this.handleAddSpell = this.handleAddSpell.bind(this)
     this.handleRemoveSpell = this.handleRemoveSpell.bind(this)
+  }
+
+  setFilter(e) {
+    this.setState({ [e.target.name]: e.target.value })
   }
 
   handleUpdate(id, charId, trait, value) {
@@ -92,24 +107,38 @@ class CharmEditor extends React.Component {
         <Typography paragraph>This Character has not yet loaded.</Typography>
       </div>
 
-    const { character, nativeCharms, martialArtsCharms, evocations, spiritCharms, spells } = this.props
+    const {
+      character, nativeCharms, abilities, martialArtsCharms, evocations,
+      spiritCharms, spells
+    } = this.props
     const {
       handleUpdate, handleRemove, handleUpdateSpell, handleRemoveSpell,
       handleAddNative, handleAddMA, handleAddEvocation, handleAddSpell, handleAddSpirit
     } = this
+    const { charmFilter } = this.state
 
     let natives = []
     let maCharms = []
     let evo = []
     let spirit = []
     let spl = []
-    natives = nativeCharms.map((c) =>
-      <Grid item xs={ 12 } md={ 6 } key={ c.id }>
-        <CharmFields charm={ c } character={ character }
-          onUpdate={ handleUpdate } onRemove={ handleRemove }
-        />
-      </Grid>
-    )
+
+    if (this.state.charmFilter === '')
+      natives = nativeCharms.map((c) =>
+        <Grid item xs={ 12 } md={ 6 } key={ c.id }>
+          <CharmFields charm={ c } character={ character }
+            onUpdate={ handleUpdate } onRemove={ handleRemove }
+          />
+        </Grid>
+      )
+    else
+      natives = nativeCharms.filter((c) => c.ability === charmFilter).map((c) =>
+        <Grid item xs={ 12 } md={ 6 } key={ c.id }>
+          <CharmFields charm={ c } character={ character }
+            onUpdate={ handleUpdate } onRemove={ handleRemove }
+          />
+        </Grid>
+      )
     maCharms = martialArtsCharms.map((c) =>
       <Grid item xs={ 12 } md={ 6 } key={ c.id }>
         <CharmFields charm={ c } character={ character }
@@ -151,6 +180,11 @@ class CharmEditor extends React.Component {
                 Add Charm&nbsp;
                 <ContentAddCircle />
               </Button>
+
+
+              <CharmFilter abilities={ abilities } filter={ charmFilter }
+                name="charmFilter" onChange={ this.setFilter }
+              />
             </Typography>
           </Grid>
           { natives }
@@ -215,6 +249,7 @@ class CharmEditor extends React.Component {
 CharmEditor.propTypes = {
   character: PropTypes.object,
   nativeCharms: PropTypes.arrayOf(PropTypes.object),
+  abilities: PropTypes.arrayOf(PropTypes.string),
   martialArtsCharms: PropTypes.arrayOf(PropTypes.object),
   evocations: PropTypes.arrayOf(PropTypes.object),
   spells: PropTypes.arrayOf(PropTypes.object),
@@ -228,38 +263,33 @@ CharmEditor.propTypes = {
 }
 
 function mapStateToProps(state, ownProps) {
-  const character = state.entities.characters[ownProps.match.params.characterId] || {}
+  const id = ownProps.match.params.characterId
+  const character = getSpecificCharacter(state, id)
 
   let nativeCharms = []
+  let abilities = []
   let martialArtsCharms = []
   let evocations = []
   let artifacts = []
-  let spells = []
   let spiritCharms = []
+  let spells = []
 
-  if (character.charms != undefined) {
-    nativeCharms = character.charms.map((id) => state.entities.charms[id])
-  }
-
-  if (character.evocations != undefined) {
-    evocations = character.evocations.map((id) => state.entities.charms[id])
-  }
-  if (character.martial_arts_charms != undefined) {
-    martialArtsCharms = character.martial_arts_charms.map((id) => state.entities.charms[id])
-  }
-  if (character.weapons != undefined) {
-    artifacts = character.merits.map((id) => state.entities.merits[id]).filter((m) => m.merit_name == 'artifact' )
-  }
-  if (character.spells != undefined) {
-    spells = character.spells.map((id) => state.entities.spells[id])
-  }
-  if (character.spirit_charms != undefined) {
-    spiritCharms = character.spirit_charms.map((id) => state.entities.charms[id])
+  if (character !== undefined) {
+    nativeCharms = getNativeCharmsForCharacter(state, id)
+    abilities = getAllAbilitiesWithCharmsForCharacter(state, id)
+    martialArtsCharms = getMartialArtsCharmsForCharacter(state, id)
+    evocations = getEvocationsForCharacter(state, id)
+    spells = getSpellsForCharacter(state, id)
+    spiritCharms = getSpiritCharmsForCharacter(state, id)
+    artifacts = getMeritsForCharacter(state, id).filter((m) =>
+      m.merit_name.toLowerCase() == 'artifact' || m.merit_name.toLowerCase() == 'hearthstone' || m.merit_name.toLowerCase() == 'warstrider'
+    )
   }
 
   return {
     character,
     nativeCharms,
+    abilities,
     martialArtsCharms,
     evocations,
     artifacts,