+++ app/javascript/lca/components/characters/charms/CharmFilter.jsx
@@ -1,0 +1,58 @@
+import React from 'react'
+import PropTypes from 'prop-types'
+
+import Button from 'material-ui/Button'
+import { MenuItem } from 'material-ui/Menu'
+import TextField from 'material-ui/TextField'
+import Grow from 'material-ui/transitions/Grow'
+import Filter from 'material-ui-icons/FilterList'
+
+class CharmFilter extends React.Component {
+  constructor(props) {
+    super(props)
+    this.state = { open: false }
+    this.toggleOpen = this.toggleOpen.bind(this)
+  }
+
+  toggleOpen() {
+    this.setState({ open: !this.state.open })
+  }
+  render() {
+    const { abilities, name, filter, onChange } = this.props
+    const options = abilities.map((abil) =>
+      abil === 'martial_arts' ?
+        <span key={abil} /> :
+        <MenuItem key={ abil } value={ abil } style={{ textTransform: 'capitalize' }}>
+          { abil }
+        </MenuItem>
+    )
+    return <React.Fragment>
+      <Button onClick={ this.toggleOpen }>
+        Filter Charms&nbsp;
+        <Filter />
+      </Button>
+      <Grow
+        in={ this.state.open }
+        style={{ transformOrigin: '0 0 0' }}
+      >
+        <TextField select style={{ minWidth: '8em', textTransform: 'capitalize' }}
+          name={ name }
+          label="Set Filter"
+          value={ filter }
+          onChange={ onChange }
+        >
+          <MenuItem value="">No Filter</MenuItem>
+          { options }
+        </TextField>
+      </Grow>
+    </React.Fragment>
+  }
+}
+CharmFilter.propTypes = {
+  abilities: PropTypes.arrayOf(PropTypes.string).isRequired,
+  name: PropTypes.string,
+  filter: PropTypes.string.isRequired,
+  onChange: PropTypes.func,
+}
+
+export default CharmFilter