+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v42
 ###### *Not yet released*
+- The Charms page can now be filtered by Ability / Attribute
 - Rating fields for Attributes and Abilities now change to the number you expect instead of 5 if the cursor is not in the most auspicious spot (thanks /u/evilmegan)
 - Merits can now be ranked N/A